댓글 삭제 구현(대댓글이 달린 상태면 삭제한 메세지입니다 표시) boardGet.jsp CommentDAO CommentService

+++ src/main/java/com/example/movie/service/CommentService.java
@@ -91,9 +91,14 @@ public enum CommentService {
         log.info("removeComment()...");
 
         int commentNo = Integer.parseInt(request.getParameter("commentNo"));
-
+        int parentNo = Integer.parseInt(request.getParameter("parentNo"));
         int contentNo = commentDAO.getContentNoByComment(commentNo);
-        boolean result = commentDAO.deleteComment(commentNo);
+        boolean result;
+        if(commentDAO.checkHasRe(parentNo) && commentNo == parentNo)
+            result = commentDAO.updateCommentDie(commentNo);
+        else
+            result = commentDAO.deleteComment(commentNo);
+
         int cnt = commentDAO.commentCnt(contentNo);
         BoardDAO boardDAO = new BoardDAO();
         try{
